Nearest Neighbors Query by Random Selection

Nearest Neighbors Query by Random Selection

Usage

random_knn_query(
  query,
  reference,
  k,
  metric = "euclidean",
  use_alt_metric = TRUE,
  order_by_distance = TRUE,
  n_threads = 0,
  verbose = FALSE,
  obs = "R"
)

Arguments

query

Matrix of n query items, with observations in the rows and features in the columns. Optionally, the data may be passed with the observations in the columns, by setting obs = "C", which should be more efficient. The reference data must be passed in the same orientation as query.

reference

Matrix of m reference items, with observations in the rows and features in the columns. The nearest neighbors to the queries are randomly selected from this data. Optionally, the data may be passed with the observations in the columns, by setting obs = "C", which should be more efficient. The query data must be passed in the same orientation as reference.

k

Number of nearest neighbors to return.

metric

Type of distance calculation to use. One of:

"euclidean".

"l2sqr" (squared Euclidean).

"cosine".

"cosine-preprocess": cosine with preprocessing: this trades memory for a potential speed up during the distance calculation.It should give the same results as cosine, give or take minor numerical changes. Be aware that the distance between two identical items may not always give exactly zero with this method.

"manhattan".

"correlation" (1 minus the Pearson correlation).

"correlation-preprocess": correlation with preprocessing. This trades memory for a potential speed up during the distance calculation. It should give the same results as correlation, give or take minor numerical changes. Be aware that the distance between two identical items may not always give exactly zero with this method.

"hamming".

"bhamming" (hamming on binary data with bitset internal memory optimization).

use_alt_metric

If TRUE, use faster metrics that maintain the ordering of distances internally (e.g. squared Euclidean distances if using metric = "euclidean"), then apply a correction at the end. Probably the only reason to set this to FALSE is if you suspect that some sort of numeric issue is occurring with your data in the alternative code path.

order_by_distance

If TRUE (the default), then results for each item are returned by increasing distance. If you don't need the results sorted, e.g. you are going to pass the results as initialization to another routine like graph_knn_query(), set this to FALSE to save a small amount of computational time.

n_threads

Number of threads to use.

verbose

If TRUE, log information to the console.

obs

set to "C" to indicate that the input query and reference orientation stores each observation as a column (the orientation must be consistent). The default "R" means that observations are stored in each row. Storing the data by row is usually more convenient, but internally your data will be converted to column storage. Passing it already column-oriented will save some memory and (a small amount of) CPU usage.

Value

an approximate nearest neighbor graph as a list containing:

idx an n by k matrix containing the nearest neighbor indices.

dist an n by k matrix containing the nearest neighbor distances.

Examples

# 100 reference iris items
iris_ref <- iris[iris$Species %in% c("setosa", "versicolor"), ]

# 50 query items
iris_query <- iris[iris$Species == "versicolor", ]

# For each item in iris_query find 4 random neighbors in iris_ref
# If you pass a data frame, non-numeric columns are removed
# set verbose = TRUE to get details on the progress being made
iris_query_random_nbrs <- random_knn_query(iris_query,
  reference = iris_ref,
  k = 4, metric = "euclidean", verbose = TRUE
)
#> 02:16:59 Generating random k-nearest neighbor graph from reference with k = 4
#> 02:16:59 Finished

# Manhattan (l1) distance
iris_query_random_nbrs <- random_knn_query(iris_query,
  reference = iris_ref,
  k = 4, metric = "manhattan"
)