Prepare Approximate Nearest Neighbors Index for Querying

Takes a nearest neighbor index produced by rnnd_build and uses the graph diversification method of Harwood and Drummond (2016) to prepare a nearest neighbor index for use in querying with rnnd_query().

Usage

rnnd_prepare(index, n_threads = 0, verbose = FALSE)

Arguments

index

An approximate nearest neighbor index produced by rnnd_build().

n_threads

Number of threads to use.

verbose

If TRUE, log information to the console.

Value

the approximate nearest neighbor index, prepared for querying.

Details

If you ran rnnd_build() with prepare = TRUE you do not need to run this function before calling rnnd_query(). If you run rnnd_query() on an index which has not been prepared, it will be prepared before the query is run.

Calling rnnd_prepare on an already prepared index does nothing.

References

Harwood, B., & Drummond, T. (2016). Fanng: Fast approximate nearest neighbour graphs. In Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition (pp. 5713-5722).

See also

Examples

iris_even <- iris[seq_len(nrow(iris)) %% 2 == 0, ]
iris_odd <- iris[seq_len(nrow(iris)) %% 2 == 1, ]

# Build index for knn only
iris_even_index <- rnnd_build(iris_even, k = 4)

# Prepare it for querying
iris_even_index <- rnnd_prepare(iris_even_index)

# Query new data
iris_odd_nbrs <- rnnd_query(index = iris_even_index, query = iris_odd, k = 4)