Merge two approximate nearest neighbors graphs

Merge two approximate nearest neighbors graphs

Usage

merge_knn(
  nn_graph1,
  nn_graph2,
  is_query = FALSE,
  n_threads = 0,
  verbose = FALSE
)

Arguments

nn_graph1

A nearest neighbor graph to merge. Should consist of a list containing:

idx an n by k matrix containing the k nearest neighbor indices.

dist an n by k matrix containing k nearest neighbor distances.

nn_graph2

Another nearest neighbor graph to merge with the same format as nn_graph1. The number of neighbors can differ between graphs, but the merged result will have the same number of neighbors as specified in nn_graph1.

is_query

If TRUE then the graphs are treated as the result of a knn query, not a knn building process. This should be set to TRUE if nn_graph1 and nn_graph2 are the results of using e.g. graph_knn_query() or random_knn_query(), and set to FALSE if these are the results of nnd_knn() or random_knn(). The difference is that if is_query = FALSE, if an index p is found in nn_graph1[i, ], i.e. p is a neighbor of i with distance d, then it is assumed that i is a neighbor of p with the same distance. If is_query = TRUE, then i and p are indexes into two different datasets and the symmetry does not hold. If you aren't sure what case applies to you, it's safe (but potentially inefficient) to set is_query = TRUE.

n_threads

Number of threads to use.

verbose

If TRUE, log information to the console.

Value

a list containing:

idx an n by k matrix containing the merged nearest neighbor indices.

dist an n by k matrix containing the merged nearest neighbor distances.

The size of k in the output graph is the same as that of nn_graph1.

Examples

set.seed(1337)
# Nearest neighbor descent with 15 neighbors for iris three times,
# starting from a different random initialization each time
iris_rnn1 <- nnd_knn(iris, k = 15, n_iters = 1)
iris_rnn2 <- nnd_knn(iris, k = 15, n_iters = 1)

# Merged results should be an improvement over either individual results
iris_mnn <- merge_knn(iris_rnn1, iris_rnn2)
sum(iris_mnn$dist) < sum(iris_rnn1$dist)
#> [1] TRUE
sum(iris_mnn$dist) < sum(iris_rnn2$dist)
#> [1] TRUE