Find Nearest Neighbors and Distances

Find Nearest Neighbors and Distances

Usage

nnd_knn(
  data,
  k = NULL,
  metric = "euclidean",
  init = NULL,
  n_iters = 10,
  max_candidates = NULL,
  delta = 0.001,
  low_memory = TRUE,
  use_alt_metric = TRUE,
  n_threads = 0,
  verbose = FALSE,
  progress = "bar",
  obs = "R"
)

Arguments

data

Matrix of n items to generate neighbors for, with observations in the rows and features in the columns. Optionally, input can be passed with observations in the columns, by setting obs = "C", which should be more efficient.

k

Number of nearest neighbors to return. Optional if init is specified.

metric

Type of distance calculation to use. One of:

"euclidean".

"l2sqr" (squared Euclidean).

"cosine".

"cosine-preprocess": cosine with preprocessing: this trades memory for a potential speed up during the distance calculation.It should give the same results as cosine, give or take minor numerical changes. Be aware that the distance between two identical items may not always give exactly zero with this method.

"manhattan".

"correlation" (1 minus the Pearson correlation).

"correlation-preprocess": correlation with preprocessing. This trades memory for a potential speed up during the distance calculation. It should give the same results as correlation, give or take minor numerical changes. Be aware that the distance between two identical items may not always give exactly zero with this method.

"hamming".

"bhamming" (hamming on binary data with bitset internal memory optimization).

init

Initial data neighbor graph to optimize. If not provided, k random neighbors are created. If provided, the input format should be a list containing:

idx an n by k matrix containing the nearest neighbor indices.

dist (optional) an n by k matrix containing the nearest neighbor distances.

If k and init are specified as arguments to this function, and the number of neighbors provided in init is not equal to k then:

if k is smaller, only the k closest values in init are retained.

if k is larger, then random neighbors will be chosen to fill init to the size of k. Note that there is no checking if any of the random neighbors are duplicates of what is already in init so effectively fewer than k neighbors may be chosen for some observations under these circumstances.

If the input distances are omitted, they will be calculated for you.

n_iters

Number of iterations of nearest neighbor descent to carry out.

max_candidates

Maximum number of candidate neighbors to try for each item in each iteration. Use relative to k to emulate the "rho" sampling parameter in the nearest neighbor descent paper. By default, this is set to k or 60, whichever is smaller.

delta

The minimum relative change in the neighbor graph allowed before early stopping. Should be a value between 0 and 1. The smaller the value, the smaller the amount of progress between iterations is allowed. Default value of 0.001 means that at least 0.1% of the neighbor graph must be updated at each iteration.

low_memory

If TRUE, use a lower memory, but more computationally expensive approach to index construction. If set to FALSE, you should see a noticeable speed improvement, especially when using a smaller number of threads, so this is worth trying if you have the memory to spare.

use_alt_metric

If TRUE, use faster metrics that maintain the ordering of distances internally (e.g. squared Euclidean distances if using metric = "euclidean"), then apply a correction at the end. Probably the only reason to set this to FALSE is if you suspect that some sort of numeric issue is occurring with your data in the alternative code path.

n_threads

Number of threads to use.

verbose

If TRUE, log information to the console.

progress

Determines the type of progress information logged if verbose = TRUE. Options are:

"bar": a simple text progress bar.

"dist": the sum of the distances in the approximate knn graph at the end of each iteration.

obs

set to "C" to indicate that the input data orientation stores each observation as a column. The default "R" means that observations are stored in each row. Storing the data by row is usually more convenient, but internally your data will be converted to column storage. Passing it already column-oriented will save some memory and (a small amount of) CPU usage.

Value

the approximate nearest neighbor graph as a list containing:

idx an n by k matrix containing the nearest neighbor indices.

dist an n by k matrix containing the nearest neighbor distances.

References

Dong, W., Moses, C., & Li, K. (2011, March). Efficient k-nearest neighbor graph construction for generic similarity measures. In Proceedings of the 20th international conference on World Wide Web (pp. 577-586). ACM. https://doi.org/10.1145/1963405.1963487.

Examples

# Find 4 (approximate) nearest neighbors using Euclidean distance
# If you pass a data frame, non-numeric columns are removed
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean")

# Manhattan (l1) distance
iris_nn <- nnd_knn(iris, k = 4, metric = "manhattan")

# Multi-threading: you can choose the number of threads to use: in real
# usage, you will want to set n_threads to at least 2
iris_nn <- nnd_knn(iris, k = 4, metric = "manhattan", n_threads = 1)

# Use verbose flag to see information about progress
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", verbose = TRUE)
#> 22:28:54 Initializing from random neighbors
#> 22:28:54 Generating random k-nearest neighbor graph with k = 4
#> 22:28:54 Finished
#> 22:28:54 Running nearest neighbor descent for 10 iterations
#> 22:28:54 Finished

# Nearest neighbor descent uses random initialization, but you can pass any
# approximation using the init argument (as long as the metrics used to
# calculate the initialization are compatible with the metric options used
# by nnd_knn).
iris_nn <- random_knn(iris, k = 4, metric = "euclidean")
iris_nn <- nnd_knn(iris, init = iris_nn, metric = "euclidean", verbose = TRUE)
#> 22:28:54 Running nearest neighbor descent for 10 iterations
#> 22:28:54 Finished

# Number of iterations controls how much optimization is attempted. A smaller
# value will run faster but give poorer results
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", n_iters = 2)

# You can also control the amount of work done within an iteration by
# setting max_candidates
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", max_candidates = 50)

# Optimization may also stop early if not much progress is being made. This
# convergence criterion can be controlled via delta. A larger value will
# stop progress earlier. The verbose flag will provide some information if
# convergence is occurring before all iterations are carried out.
set.seed(1337)
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", n_iters = 5, delta = 0.5)

# To ensure that descent only stops if no improvements are made, set delta = 0
set.seed(1337)
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", n_iters = 5, delta = 0)

# A faster version of the algorithm is available that avoids repeated
# distance calculations at the cost of using more RAM. Set low_memory to
# FALSE to try it.
set.seed(1337)
iris_nn <- nnd_knn(iris, k = 4, metric = "euclidean", low_memory = FALSE)